# Convert process_add_dir to async