--- a/Cargo.toml
+++ b/Cargo.toml
@@ -88,6 +88,7 @@ minus = { version = "5.4.0", features = ["static_output", "search"] }
 nom = "7.1.3"
 num_cpus = "1.16.0"
 parking_lot = "0.12.1"
+par-stream = { version = "0.10.2", features = ["runtime-tokio"] }
 percent-encoding = "2.1"
 pluralizer = "0.4.0"
 polars = { version = "0.45.0", features = [
@@ -134,6 +135,7 @@ tempfile = "3.8.0"
 threadpool = "1.8.1"
 time = { version = "0.3.28", features = ["serde"] }
 tokio = { version = "1.32.0", features = ["full"] }
+tokio-stream = "0.1.17"
 tokio-util = "0.7.8"
 toml = "0.8.19"
 url = "2.4.1"