<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

* **Refactor**
  * Enhanced file addition with fully asynchronous, concurrent directory traversal and processing for improved performance.
  * Improved progress reporting accuracy during file additions.

* **Chores**
  * Removed deprecated synchronous file storage methods from storage components.
  * Added new dependencies supporting asynchronous streaming and concurrency.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->