@Lilianxr 2025-07-24
Performance Comparison with current main:
| Add Benchmark (ms) | Main   | Async-add | Δ (%)  |
|--------------------|--------|-----------|--------|
| 10k                | 14.44  | 15.28     | +5.78%  |
| 100k               | 56.35  | 62.33     | +10.61% |
| 1000k              | 444.99 | 471.01    | +5.85%  |